backend/modules/plagiarism/parapahsedetection/requirements.txt
flask
flask-cors
sentence-transformers
scikit-learn
numpy
joblib
mysql-connector-python
requests
ddgs
trafilatura
playwright
urllib3
sinling
